requirements.txt
# Core ML / DL
tensorflow==2.16.1
keras==3.3.3
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2

# Computer Vision
opencv-python==4.10.0.84
mtcnn==0.1.1
Pillow==10.3.0

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Explainability
grad-cam==1.5.5

# Web App
streamlit==1.38.0

# Database / Cloud
sqlalchemy==2.0.32
supabase==2.4.6
python-dotenv==1.0.1

# Utils
requests==2.32.3
tqdm==4.66.5

# Testing
pytest==8.3.3
